api/pyproject.toml
[project]
name = "fairshare"
version = "0.1.0"
description = "Open-source expense splitting and settlement tracking app. Split bills fairly, track shared expenses, and settle up with friends."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.17.2",
    "email-validator==2.3.0",
    "fastapi==0.128.0",
    "psycopg[binary]==3.3.2",
    "pydantic-settings==2.12.0",
    "pyjwt==2.10.1",
    "pwdlib[argon2]==0.3.0",
    "python-multipart==0.0.21",
    "sqlmodel==0.0.31",
    "uvicorn==0.40.0",
]

[dependency-groups]
dev = [
    "ruff==0.14.10",
    "fastapi-cli==0.0.20",
    "ty==0.0.10",
    "pytest==8.4.1",
    "pytest-asyncio==1.0.0",
    "httpx==0.28.1",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.pyproject-fmt]
keep_full_version = true
